testing
I'll be adding a couple pieces of my work to the page in a little bit and plan to host them here with a link to here from the academics page. Right now it's just a matter of figuring out how this page looks and then going to town uploading files from there. Markdown might be easier for this. We shall see.